requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
xgboost==2.1.1
lightgbm==4.5.0
matplotlib==3.9.2
seaborn==0.13.2
flask==3.0.3
pyyaml==6.0.1